Move javascript helper methods to the test-utils
evaluateFail and evaluateJavascript
Simplify Ifs logic for number of parameters expected
@armanshan12
